+++ src/Components/Button/_styles.js
@@ -1,4 +1,4 @@
-const styles = {
+const styles = (theme) => ({
     root: {
         background: 'linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%)',
         border: 0,
@@ -6,8 +6,8 @@ const styles = {
         boxShadow: '0 3px 5px 2px rgba(255, 105, 135, .3)',
         color: 'white',
         height: 48,
-        padding: '0 30px',
+        padding: theme.spacing(0, 4),
     },
-};
+});
 
 export default styles;
