+++ src/Pages/Main/_styles.js
@@ -1,0 +1,10 @@
+
+const styles = (theme) => ({
+    root: {
+        width: '100%',
+        maxWidth: 360,
+        backgroundColor: theme.palette.background.paper,
+    },
+});
+
+export default styles;
