+++ src/Pages/Tables/_styles.js
@@ -1,0 +1,12 @@
+const styles = (theme) => ({
+    root: {
+        flexGrow: 1,
+        backgroundColor: theme.palette.grey.A500,
+        overflow: 'hidden',
+        marginTop: theme.spacing(3),
+        padding: theme.spacing(3),
+    },
+});
+
+
+export default styles;
